remove: libft submodule

+++ includes/mesh.h
@@ -33,13 +33,14 @@ typedef struct s_triangle
 typedef struct s_mesh
 {
 	t_point3_vct	vertices;
+
 }	t_mesh;
 
-int8_t		new_mesh(t_mesh	*mesh);
-size_t		mesh_tri_count(t_mesh *mesh);
-int8_t		delete_mesh(t_mesh *mesh);
-int8_t		add_triangle(t_mesh *mesh, t_triangle *triangle);
-t_triangle	new_triangle(t_point3 v1, t_point3 v2, t_point3 v3);
-t_triangle	*get_triangle(t_mesh *mesh, size_t i);
+int8_t			new_mesh(t_mesh	*mesh);
+uint32_t		mesh_tri_count(t_mesh *mesh);
+int8_t			delete_mesh(t_mesh *mesh);
+int8_t			add_triangle(t_mesh *mesh, t_triangle *triangle);
+t_triangle		new_triangle(t_point3 v1, t_point3 v2, t_point3 v3);
+t_triangle		*get_triangle(t_mesh *mesh, uint32_t i);
 
 #endif
